Added "Question" and "Answer" where questions contain 4 the question (String) answers (List), and answer contains an answer (String) and boolean to indicate if the answer is corrrect.
Added finally clause to serverlistener to properly close the serversocket once the loop is ended.

Added player instances to the playerResponse function in GameProtocol to properly account for points gained if the correct answer is given.

Changed the serverSideGame to use instanceof to check if an answer (Boolean) is sent from the client, this should provide more flexibility for the client to send stuff other than answers to questions.